added common component coupopncode
