+++ client/src/components/Common/CouponCode/CouponCode.style.js
@@ -1,0 +1,14 @@
+/* eslint-disable import/prefer-default-export */
+import styled from 'styled-components';
+import { colors, borders, size } from '../../../theme';
+
+export const CouponInputWrapper = styled.div`
+  // width: 200px;
+
+  // border: ${borders.newInputBox};
+  // border-radius: 10px;
+
+  // @media (max-width: ${size.mobileL}) {
+  //   width: 160px;
+  // }
+`;
